refactor: extract getNameParamNode util
