+++ lib/rules/no-unused-emit-declarations.js
@@ -9,6 +9,7 @@ const utils = require('../utils')
 
 /**
  * @typedef {import('../utils').ComponentEmit} ComponentEmit
+ * @typedef {import('../utils').NameWithLoc} NameWithLoc
  * @typedef {import('../utils').VueObjectData} VueObjectData
  */
 
@@ -17,30 +18,6 @@ const utils = require('../utils')
  * @property {Set<Identifier>} contextReferenceIds
  * @property {Set<Identifier>} emitReferenceIds
  */
-
-/**
- * @typedef {object} NameWithLoc
- * @property {string} name
- * @property {SourceLocation} loc
- * @property {Range} range
- */
-
-/**
- * Get the name param node from the given CallExpression
- * @param {CallExpression} node CallExpression
- * @returns {NameWithLoc | null}
- */
-function getNameParamNode(node) {
-  const nameLiteralNode = node.arguments[0]
-  if (nameLiteralNode && utils.isStringLiteral(nameLiteralNode)) {
-    const name = utils.getStringLiteralValue(nameLiteralNode)
-    if (name != null) {
-      return { name, loc: nameLiteralNode.loc, range: nameLiteralNode.range }
-    }
-  }
-
-  return null
-}
 
 /**
  * Check if the given node is a reference of setup context
@@ -85,7 +62,7 @@ module.exports = {
      * @param {CallExpression} node
      * */
     function addEmitCall(node) {
-      const nameParamNode = getNameParamNode(node)
+      const nameParamNode = utils.getNameParamNode(node)
       if (nameParamNode) {
         emitCalls.set(nameParamNode.name, node)
       }
