+++ lib/utils/index.js
@@ -26,6 +26,7 @@ const { getScope } = require('./scope')
  * @typedef {import('../../typings/eslint-plugin-vue/util-types/utils').ComponentInferTypeEmit} ComponentInferTypeEmit
  * @typedef {import('../../typings/eslint-plugin-vue/util-types/utils').ComponentUnknownEmit} ComponentUnknownEmit
  * @typedef {import('../../typings/eslint-plugin-vue/util-types/utils').ComponentEmit} ComponentEmit
+ * @typedef {import('../../typings/eslint-plugin-vue/util-types/utils').NameWithLoc} NameWithLoc
  * @typedef {import('../../typings/eslint-plugin-vue/util-types/utils').ComponentTypeSlot} ComponentTypeSlot
  * @typedef {import('../../typings/eslint-plugin-vue/util-types/utils').ComponentInferTypeSlot} ComponentInferTypeSlot
  * @typedef {import('../../typings/eslint-plugin-vue/util-types/utils').ComponentUnknownSlot} ComponentUnknownSlot
@@ -1087,6 +1088,12 @@ module.exports = {
    * @return {string|null} The string if static. Otherwise, null.
    */
   getStringLiteralValue,
+  /**
+   * Gets the name parameter from the given call expression.
+   * @param {CallExpression} node - The call expression to get.
+   * @returns {NameWithLoc|null} The name and location if static.
+   */
+  getNameParamNode,
   /**
    * Get all props by looking at all component's properties
    * @param {ObjectExpression} componentObject Object with component definition
@@ -2715,6 +2722,22 @@ function getStringLiteralValue(node, stringOnly) {
     node.quasis.length === 1
   ) {
     return node.quasis[0].value.cooked
+  }
+  return null
+}
+
+/**
+ * Gets the name parameter from the given call expression.
+ * @param {CallExpression} node - The call expression to get.
+ * @returns {NameWithLoc|null} The name and location if static.
+ */
+function getNameParamNode(node) {
+  const nameLiteralNode = node.arguments[0]
+  if (nameLiteralNode && isStringLiteral(nameLiteralNode)) {
+    const name = getStringLiteralValue(nameLiteralNode)
+    if (name != null) {
+      return { name, loc: nameLiteralNode.loc, range: nameLiteralNode.range }
+    }
   }
   return null
 }
